BrainTV-Brain_Thought_Visualization--main/neel_tmp/demo.py
import time
import nvtx


@nvtx.annotate(color="blue")
def my_function():
    for i in range(5):
        with nvtx.annotate("my_loop", color="red"):
            time.sleep(i)


my_function()
print("Neel job is done")
