BrainTV-Brain_Thought_Visualization--main/requirements/python311_requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.12.4
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lpips==0.1.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==23.2
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.0.9.post0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.3.3
scikit-image==0.22.0
scipy==1.11.3
sentry-sdk==1.31.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
taming-transformers==0.0.1
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.14.1
torch==2.1.0
torch-fidelity==0.3.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.0.6
wandb==0.15.12
yarl==1.9.2
